Change type decimal cords to float pg

+++ src/database/migrations/00_create_points.js
@@ -16,8 +16,8 @@ exports.up = async function up(knex) {
         table.string('zip_code').notNullable();
         table.integer('addressNumber').notNullable();
         table.string('neighborhood').notNullable();
-        table.decimal('longitude').notNullable();
-        table.decimal('latitude').notNullable();
+        table.float('longitude').notNullable();
+        table.float('latitude').notNullable();
     });
 }
